+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support file with evidence but no copyright or license
 - Encoding fix for copyrights
 - Improved debugging
+- LicenseRef values
 
 ## [2.0.3] - 2022-10-03
 ### Fixed